+++ src/asgi_user_agents/contrib/django/__init__.py
@@ -1,0 +1,8 @@
+"""Django integration for asgi-user-agents."""
+
+from asgi_user_agents.contrib.django.middleware import (
+    UserAgentMiddleware as UserAgentMiddleware,
+)
+from asgi_user_agents.contrib.django.utils import (
+    get_user_agent as get_user_agent,
+)